Split lease payment columns into two files. Moved logic about variable payment to a class. Fixed issue with totals not beign calculated.
